% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_active_collab_token}
\alias{get_active_collab_token}
\title{Get an active collab token}
\usage{
get_active_collab_token(email = "colin@thinkr.fr",
  password = rstudioapi::askForPassword())
}
\arguments{
\item{email}{your email}

\item{password}{your password. If in RStudio, it opens a connexion windows.}
}
\value{
a list of class `collab_token`
}
\description{
Get an active collab token
}
