% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_projects_list}
\alias{get_projects_list}
\title{Get the list of project}
\usage{
get_projects_list(token)
}
\arguments{
\item{token}{a token of class `collab_token`}
}
\value{
a tibble
}
\description{
Get the list of project
}
\examples{
\dontrun{
get_projects_list(my_token)
}
}
